+++ src/universal/decoys/LinkHeader.mjs
@@ -1,7 +1,7 @@
 /**
  * An empty export used by [`withGraphQLConfig`]{@link withGraphQLConfig} to
- * alias the server only module `next-graphql-react/server/mergeLinkHeaders` for
+ * alias the server only module `next-graphql-react/server/LinkHeader` for
  * a client bundle.
  * @ignore
  */
-export let mergeLinkHeaders
+export let LinkHeader