+++ src/universal/decoys/filterLinkHeader.mjs
@@ -1,7 +1,0 @@
-/**
- * An empty export used by [`withGraphQLConfig`]{@link withGraphQLConfig} to
- * alias the server only module `next-graphql-react/server/filterLinkHeader` for
- * a client bundle.
- * @ignore
- */
-export let filterLinkHeader