+++ src/server/withGraphQLConfig.mjs
@@ -1,8 +1,8 @@
 /**
  * A higher-order function to decorate a Next.js custom config in
  * `next.config.js` for [`graphql-react`](https://npm.im/graphql-react), that
- * excludes server only `graphql-react/server` and `next-graphql-react/server`
- * imports from the client bundle.
+ * excludes server only `graphql-react/server` and
+ * `next-graphql-react/server/LinkHeader` imports from the client bundle.
  * @see [Next.js custom config docs](https://nextjs.org/docs/#custom-configuration).
  * @see [`graphql-react` `ssr` docs](https://github.com/jaydenseric/graphql-react#function-ssr).
  * @kind function