+++ src/universal/withGraphQLApp.mjs
@@ -12,6 +12,12 @@ import React from 'react'
  * in `pages/_app.js` for [`graphql-react`](https://npm.im/graphql-react),
  * enabling descendant GraphQL operations with server side rendering and client
  * side data hydration.
+ *
+ * It also forwards HTTP
+ * [`Link`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Link)
+ * `rel="preload"` headers from GraphQL responses during SSR in the Next.js page
+ * response. Link URLs are forwarded unmodified, so avoid sending relative URLs
+ * from a GraphQL server hosted on a different domain to the app.
  * @see [Next.js custom `App` docs](https://nextjs.org/docs#custom-app).
  * @see [React higher-order component docs](https://reactjs.org/docs/higher-order-components).
  * @kind function