+++ changelog.md
@@ -5,6 +5,7 @@
 ### Patch
 
 - Fixed edge case HTTP `Link` header parsing bugs (e.g. an error when a URL contains `,`) by replacing the `filterLinkHeader` and `mergeLinkHeaders` functions with [`http-link-header`](https://npm.im/http-link-header).
+- Fixed `withGraphQLConfig` documentation.
 
 ## 3.1.0